Fix clang-12 -Wimplicit-const-int-float-conversion warnings
This fix may not be perfectly correct, but it silences the warnings.
:P

    hal/utils/scope_trig.c:157:17: error: implicit conversion from 'long' to 'double' changes value from 9223372036854775807 to 9223372036854775808 [-Werror,-Wimplicit-const-int-float-conversion]
            if (fp_level > INT64_MAX) {
                         ~ ^~~~~~~~~
    /usr/include/stdint.h:124:22: note: expanded from macro 'INT64_MAX'
    # define INT64_MAX              (__INT64_C(9223372036854775807))
                                     ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
    /usr/include/stdint.h:106:24: note: expanded from macro '__INT64_C'
    #  define __INT64_C(c)  c ## L
                            ^~~~~~
    <scratch space>:48:1: note: expanded from here
    9223372036854775807L
    ^~~~~~~~~~~~~~~~~~~~
    hal/utils/scope_trig.c:158:17: error: implicit conversion from 'long' to 'double' changes value from 9223372036854775807 to 9223372036854775808 [-Werror,-Wimplicit-const-int-float-conversion]
                fp_level = INT64_MAX;
                         ~ ^~~~~~~~~
    /usr/include/stdint.h:124:22: note: expanded from macro 'INT64_MAX'
    # define INT64_MAX              (__INT64_C(9223372036854775807))
                                     ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
    /usr/include/stdint.h:106:24: note: expanded from macro '__INT64_C'
    #  define __INT64_C(c)  c ## L
                            ^~~~~~
    <scratch space>:49:1: note: expanded from here
    9223372036854775807L
    ^~~~~~~~~~~~~~~~~~~~
    hal/utils/scope_trig.c:166:17: error: implicit conversion from 'unsigned long' to 'double' changes value from 18446744073709551615 to 18446744073709551616 [-Werror,-Wimplicit-const-int-float-conversion]
            if (fp_level > UINT64_MAX) {
                         ~ ^~~~~~~~~~
    /usr/include/stdint.h:130:23: note: expanded from macro 'UINT64_MAX'
    # define UINT64_MAX             (__UINT64_C(18446744073709551615))
                                     ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
    /usr/include/stdint.h:107:25: note: expanded from macro '__UINT64_C'
    #  define __UINT64_C(c) c ## UL
                            ^~~~~~~
    <scratch space>:52:1: note: expanded from here
    18446744073709551615UL
    ^~~~~~~~~~~~~~~~~~~~~~
    hal/utils/scope_trig.c:167:17: error: implicit conversion from 'unsigned long' to 'double' changes value from 18446744073709551615 to 18446744073709551616 [-Werror,-Wimplicit-const-int-float-conversion]
                fp_level = UINT64_MAX;
                         ~ ^~~~~~~~~~
    /usr/include/stdint.h:130:23: note: expanded from macro 'UINT64_MAX'
    # define UINT64_MAX             (__UINT64_C(18446744073709551615))
                                     ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
    /usr/include/stdint.h:107:25: note: expanded from macro '__UINT64_C'
    #  define __UINT64_C(c) c ## UL
                            ^~~~~~~
    <scratch space>:53:1: note: expanded from here
    18446744073709551615UL
    ^~~~~~~~~~~~~~~~~~~~~~
    4 errors generated.
    make: *** [Makefile:378: objects/hal/utils/scope_trig.o] Error 1

+++ src/hal/utils/scope_trig.c
@@ -154,8 +154,8 @@ void refresh_trigger(void)
 	break;
 
     case HAL_S64:
-	if (fp_level > INT64_MAX) {
-	    fp_level = INT64_MAX;
+	if (fp_level > (double)INT64_MAX) {
+	    fp_level = (double)INT64_MAX;
 	}
 	if (fp_level < INT64_MIN) {
 	    fp_level = INT64_MIN;
@@ -163,8 +163,8 @@ void refresh_trigger(void)
 	ctrl_shm->trig_level.d_s64 = fp_level;
 	break;
     case HAL_U64:
-	if (fp_level > UINT64_MAX) {
-	    fp_level = UINT64_MAX;
+	if (fp_level > (double)UINT64_MAX) {
+	    fp_level = (double)UINT64_MAX;
 	}
 	if (fp_level < 0.0) {
 	    fp_level = 0.0;
